rename package s/firehose/writer
to avoid namespace conflicts with aws's firehose pkg
